Add files via upload
supply file named ips.txt with ip on each line. This file needs to be in same path as that of batch script or full file path needs to be specified by editing the script.
Live ips will be outputted to up-ips.txt in same directory as that of batch script.